+++ src/main/java/tech/artadevs/finances/dtos/FinancialTransactionRequestDto.java
@@ -1,5 +1,7 @@
 package tech.artadevs.finances.dtos;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.Data;
@@ -8,7 +10,7 @@ import lombok.Data;
 public class FinancialTransactionRequestDto {
 
 	@NotNull(message = "The transaction value is required.")
-	private Double value;
+	private BigDecimal value;
 
 	@Size(max = 100, message = "The length of description must be at most 100 characters.")
 	private String description;
@@ -16,7 +18,7 @@ public class FinancialTransactionRequestDto {
 	public FinancialTransactionRequestDto() {
 	}
 
-	public Double getValue() {
+	public BigDecimal getValue() {
 		return value;
 	}
 
@@ -24,7 +26,7 @@ public class FinancialTransactionRequestDto {
 		return description;
 	}
 
-	public FinancialTransactionRequestDto setValue(Double value) {
+	public FinancialTransactionRequestDto setValue(BigDecimal value) {
 		this.value = value;
 		return this;
 	}
